telegram_bot/requirements.txt
PyTelegramBotApi
telebot-dialogue
requests
beautifulsoup4
python-dotenv
lxml
psycopg2-binary